Create domain and networking modules
