--- app/src/main/java/com/jj/templateproject/data/Result.kt
@@ -1,6 +1,0 @@
-package com.jj.templateproject.data
-
-sealed class Result<out T : Any> {
-    data class Success<out T : Any>(val data: T) : Result<T>()
-    data class Error(val exception: Exception) : Result<Nothing>()
-}
