+++ app/src/main/java/com/jj/templateproject/di/koin/KoinLauncher.kt
@@ -1,6 +1,7 @@
 package com.jj.templateproject.di.koin
 
 import android.content.Context
+import com.jj.templateproject.di.networkingModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
 
@@ -9,7 +10,7 @@ class KoinLauncher {
     fun startKoin(applicationContext: Context) {
         startKoin {
             androidContext(applicationContext)
-            modules(mainModule)
+            modules(mainModule, networkingModule)
         }
     }
 }
