+++ domain/src/main/java/com/jj/templateproject/domain/BaseResult.kt
@@ -1,0 +1,6 @@
+package com.jj.templateproject.domain
+
+sealed class BaseResult<out T : Any> {
+    data class Success<out T : Any>(val data: T) : BaseResult<T>()
+    data class Error(val exception: Exception) : BaseResult<Nothing>()
+}
