+++ domain/src/main/java/com/jj/templateproject/domain/google/TemplateRepository.kt
@@ -1,0 +1,8 @@
+package com.jj.templateproject.domain.google
+
+import com.jj.templateproject.domain.BaseResult
+
+interface TemplateRepository {
+    suspend fun getGoogleData(): BaseResult<String>
+    suspend fun getGoogleStatus(): BaseResult<Unit>
+}
