+++ networking/src/main/java/com/jj/templateproject/data/google/service/TemplateService.kt
@@ -1,4 +1,4 @@
-package com.jj.templateproject.data.network.service
+package com.jj.templateproject.data.google.service
 
 import retrofit2.Response
 import retrofit2.http.GET
